+++ .gitignore
@@ -1,1 +1,1 @@
-RubyBlockConverter.pyc
+*.pyc
